refactor GroupAssociationRequest-related forms to share a common base class -- but implement it library-style (downstream form calls methods on base class) to keep things simple
